requirements.txt
requirements.txt

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Google Cloud
google-cloud-bigquery==3.13.0
google-oauth2-tool==0.0.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Google Sheets
gspread==5.12.0
gspread-dataframe==3.3.1

# Excel Support
openpyxl==3.1.2

# Development
python-dotenv==1.0.0